requirements.txt
affine==2.3.0
attrs==21.2.0
aws-cfn-bootstrap==2.0
beautifulsoup4==4.9.3
boto==2.49.0
boto3==1.18.43
botocore==1.21.43
certifi==2021.5.30
click==8.0.1
click-plugins==1.1.1
cligj==0.7.2
docutils==0.14
importlib-metadata==4.8.1
jmespath==0.10.0
joblib==1.0.1
lockfile==0.11.0
lxml==4.6.3
mysqlclient==1.4.2
nltk==3.6.2
nose==1.3.4
numpy==1.16.5
py-dateutil==2.2
pyparsing==2.4.7
pystache==0.5.4
python-daemon==2.2.3
python-dateutil==2.8.2
python37-sagemaker-pyspark==1.4.1
pytz==2021.1
PyYAML==5.4.1
rasterio==1.2.8
regex==2021.8.3
s3transfer==0.5.0
simplejson==3.2.0
six==1.13.0
snuggs==1.4.7
tqdm==4.62.1
typing-extensions==3.10.0.2
urllib3==1.26.6
windmill==1.6
zipp==3.5.0